requirements.txt
websockets==13.1
ocpp==0.26.0
aioconsole
pyyaml
flask
requests
click
cryptography
netifaces
dnslib
dnspython
dpkt
pyopenssl
wheel
setuptools
git+https://github.com/mininet/mininet@2.3.0
pillow
mako
mitmproxy
pandas
matplotlib
numpy